unified speech handling
