+++ src/exercises/recognize/BottomFeedback.js
@@ -1,24 +1,16 @@
 import { useState } from "react";
-import Speech from "speak-tts";
 import strings from "../../i18n/definitions";
+import ZeeguuSpeech from "../../speech/ZeeguuSpeech";
 
-export default function BottomFeedback({ bookmarkToStudy, correctAnswer }) {
-  const [speech] = useState(new Speech());
-
-  speech
-    .init()
-    .then((data) => {
-      // The "data" object contains the list of available voices and the voice synthesis params
-      let randomVoice = _getRandomVoice(data.voices, bookmarkToStudy.from_lang);
-
-      speech.setVoice(randomVoice.name);
-    })
-    .catch((e) => {
-      console.error("An error occured while initializing : ", e);
-    });
+export default function BottomFeedback({
+  bookmarkToStudy,
+  correctAnswer,
+  api,
+}) {
+  const [speech] = useState(new ZeeguuSpeech(api, bookmarkToStudy.from_lang));
 
   function handleSpeak() {
-    speech.speak({ text: bookmarkToStudy.from });
+    speech.speakOut(bookmarkToStudy.from);
   }
 
   return (
@@ -30,13 +22,4 @@ export default function BottomFeedback({ bookmarkToStudy, correctAnswer }) {
       </button>
     </div>
   );
-}
-
-function _randomElement(x) {
-  return x[Math.floor(Math.random() * x.length)];
-}
-
-function _getRandomVoice(voices, language) {
-  let x = _randomElement(voices.filter((v) => v.lang.includes(language)));
-  return x;
 }
